voice/requirements.txt
AWSIoTPythonSDK
gpiozero
google-cloud-speech